services/agent/requirements.txt
fastapi
uvicorn
pydantic-settings
python-dotenv
langgraph
langchain
langchain-google-genai
supabase
redis
qdrant-client
google-genai
